refactor: resolve #879 - split IdePage.vue and Screen.vue approaching 500-line limit
Extract composables from both files by responsibility:
- useIdePageState.ts (286 lines): E2E Lite fallbacks + state aggregation + screen context
- useIdeGlobalHotkeys.ts (76 lines): global keyboard shortcut handling
- useScreenRenderPipeline.ts (416 lines): Konva render lifecycle, watchers, animation

IdePage.vue: 498→343 lines, Screen.vue: 496→103 lines. Pure refactor, no behavior changes.

Closes #879
